fix: add blocking/control functions to grouping header
