+++ workflow/rules/dba.smk
@@ -5,7 +5,7 @@ import json
 from os.path import join
 from scripts.common import allocated, mk_dir_if_not_exist
 from scripts.peakcall import outputIDR, zip_peak_files, calc_effective_genome_fraction, get_manorm_sizes
-from scripts.blocking import test_for_block
+from scripts.grouping import test_for_block
 
 
 # ~~ workflow configuration
